Implement $shutdown
The owner of the bot can now stop the bot with this command. Stopping
the bot in this way will unload extensions first, and will allow for
reliable shutdown behavior. This paves the way for all the extensions
that save data to files not having to open and close the file every
time they need to reference or modify its value, which will save
computing power.