Change logs to use logging module, fix $shutdown
Now instead of chat messages being relayed with print(), they are put
to stderr through logging at the info level. The language used around
these methods and in the comments that describe them have been amended
to reflect this change.

Also, $shutdown has been revised so it lets the channel and the log know when it has
been successfully invoked.