Stop dad joke from triggering on long names
Watching dad jokes trigger on super long statements that simply start with "I'm" isn't very entertaining. It doesn't really make a lot of sense. In an attempt to remedy this, I have made it so the bot will no longer trigger on names that end up being over 20 characters long, including spaces. This may be subject to change as I fine tune the character limit.