Implement $reload, add cleanup on guild leaving
The bot now loads only extensions listed in extensions.json, which will be created on first run. glasssettings will be in extensions.json by default to serve as an example of the format. $reload is a new command that will unload all extensions in the bot and load extensions again as listed in extensions.json.

There is now code in each relevant extension to clean up its server data when the bot leaves a guild to prevent unnecessary data from accumulating.

Finally, the logging of individual messages has been reduced to DEBUG level instead of INFO.